refactor(sidebar): drop unused skill-chip helper, rename to skillTargetGroups
